Move AlignedToMoveBaseGoal
